fix: resolve lint errors - split VolatilityResult to own file, fix signal naming, extract applyVolatility function
