+++ apps/dms-material/src/app/global/global-universe/services/volatility-result.interface.ts
@@ -1,0 +1,5 @@
+export interface VolatilityResult {
+  symbol: string;
+  volatility1yr: 'decreasing' | 'increasing' | 'steady' | 'volatile' | null;
+  volatility5yr: 'decreasing' | 'increasing' | 'steady' | 'volatile' | null;
+}
